+++ CHANGELOG.md
@@ -44,6 +44,7 @@ IMPROVEMENTS:
   - commands/destroy: Exit codes changes. 0 means everything succeeded.
       1 means everything was declined. 2 means some were declined. [GH-811]
   - commands/destroy: Doesn't require box to exist anymore. [GH-1629]
+  - commands/init: force flag. [GH-3564]
   - commands/rsync-auto: Picks up and syncs provisioner folders if
       provisioners are backed by rsync.
   - commands/rsync-auto: Detects when new synced folders were added and warns